requirements.txt
astroid==1.6.5
Django==1.10
django-cors-headers==2.4.0
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
pylint==1.9.2
six==1.11.0
wrapt==1.10.11
